+++ crates/wp-oml/src/parser/oml_aggregate.rs
@@ -19,7 +19,7 @@ use crate::parser::fmt_prm::oml_aga_fmt;
 use crate::parser::fun_prm::oml_gw_fun;
 use crate::parser::keyword::{kw_crate_symbol, kw_in, kw_keys, kw_option, kw_read, kw_take};
 use crate::parser::lookup_prm::oml_aga_lookup_nocase;
-use crate::parser::map_prm::oml_aga_map;
+use crate::parser::map_prm::{oml_aga_map, oml_aga_obj_array};
 use crate::parser::match_prm::oml_aga_match;
 use crate::parser::pipe_prm; // for oml_aga_pipe_noprefix
 use crate::parser::pipe_prm::oml_aga_pipe;
@@ -139,6 +139,7 @@ pub fn oml_aggregate(data: &mut &str) -> WResult<EvalExp> {
             "match" => oml_aga_match.parse_next(data)?,
             "lookup_nocase" => oml_aga_lookup_nocase.parse_next(data)?,
             "object" => oml_aga_map.parse_next(data)?,
+            "array" => oml_aga_obj_array.parse_next(data)?,
             "pipe" => oml_aga_pipe.parse_next(data)?,
             "collect" => oml_aga_collect.parse_next(data)?,
             //"query" => oml_aga_shmlib.parse_next(data)?,